Backend vCenter: Upload ISO only with linuxkit push
